# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Amina API
  version: 1.0.0
  description: |
    Image generation and utilities API for the Amina Discord Bot.

    ## Authentication

    All `/v1/*` endpoints require API key authentication via Bearer token.

    ```bash
    curl -H "Authorization: Bearer amina_YOUR_API_KEY" https://api.4mina.app/v1/images/rank-card?...
    ```

    Get your API key at [api.4mina.app/dashboard](https://api.4mina.app/dashboard)

    ## Rate Limiting

    - **60 requests per minute** per API key
    - Rate limit headers included in responses:
      - `X-RateLimit-Limit`: Maximum requests per window
      - `X-RateLimit-Remaining`: Requests remaining
      - `X-RateLimit-Reset`: Unix timestamp when limit resets

    ## Response Format

    All JSON responses follow this structure:

    ```json
    {
      "success": true,
      "data": { ... },
      "meta": {
        "generatedAt": "2024-01-01T00:00:00.000Z"
      }
    }
    ```

    Image endpoints return SVG directly with `Content-Type: image/svg+xml`.

  contact:
    name: Amina Support
    url: https://discord.gg/4mina

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.4mina.app
    description: Production
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key from dashboard
  parameters:
    imageUrl:
      name: image
      in: query
      required: true
      schema:
        type: string
        format: uri
      description: URL of the image to process
    width:
      name: width
      in: query
      schema:
        type: integer
        default: 512
        minimum: 1
        maximum: 1024
      description: Output width in pixels
    height:
      name: height
      in: query
      schema:
        type: integer
        default: 512
        minimum: 1
        maximum: 1024
      description: Output height in pixels
  responses:
    SvgImage:
      description: SVG image
      content:
        image/svg+xml:
          schema:
            type: string
...
paths:
...
  /v1/images/overlays/approved:
    get:
      tags: [Overlays]
      summary: Add "APPROVED" stamp
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/overlays/rejected:
    get:
      tags: [Overlays]
      summary: Add "REJECTED" stamp
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/overlays/wasted:
    get:
      tags: [Overlays]
      summary: GTA "WASTED" effect
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/overlays/triggered:
    get:
      tags: [Overlays]
      summary: '"TRIGGERED" meme effect'
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'
  /v1/images/overlays/jail:
    get:
      tags: [Overlays]
      summary: Prison bars overlay
      parameters:
        - $ref: '#/components/parameters/imageUrl'
        - $ref: '#/components/parameters/width'
        - $ref: '#/components/parameters/height'
      responses:
        '200':
          $ref: '#/components/responses/SvgImage'

  # ========== Generators ==========
...